Allow safe mode bot to respond to commands.
